==> XenograftProteinProfiling-master/XenograftProteinProfiling-master/2_CompileProcessedRingData.R <==
GetName <- function(){
        # get the filename from the current working directory
        directory <- basename(getwd())
        
        # directory naming from MRR: "CHIPNAME_gaskGASETTYPE_DATE"
        # extracts and returns GASKETTYPE from directory name
        name <- unlist(strsplit(directory, split = "_"))
        name <- name[2]
        
        # define name as global variable for use in other functions
        name <<- gsub('gask','',name) # removes "gask" from name
}

CompileData <- function(){
        # load relevant libraries
        library(tidyverse)
        
        # get working directory to reset at end of function
        directory <- getwd()
        
        # generate list of runs and empty data frame
        listData <- list.dirs(recursive = FALSE) # gets list for top level dirs
        df <- data.frame()
        
        # iterates through each directory (i.e., run)
        for (i in 1:length(listData)){
                tempDir <- listData[i]
                setwd(tempDir)
                names <- GetName()
                listFiles <- list.files(recursive = TRUE)
                netFiles <- grep("net", listFiles)
                for( i in 1:length(netFiles)){
                        dat <- read_csv(listFiles[netFiles[i]])      
                        dat$run <- names # adds run name to data frame
                        df <- rbind(df, dat) # combines runs into single data frame
                }
                setwd(directory) #returns working dir to starting dir
        }
        # saves data
        write_csv(df, 'compiledData.csv')
}

CleanData <- function(filename = 'runs.csv'){
        # load relevant libraries
        library(tidyverse)
        
        # gets run information, this file was manually created
        dat.info <- read_csv(filename)
        
        # generates list of treatments and removes NA data and blank runs
        treatments <- unique(dat.info$Treatment, na.rm = TRUE)
        remove <- c(NA, 'NA', 'Blank')    
        treatments <- treatments[!treatments %in% remove]
        
        # gets compiled data generated by CompileData function
        dat <- read_csv('compiledData.csv')
        
        # creates empty data frame to store labeled data
        df <- data.frame()
        
        # iterates through each row in compiled data
        # assigns to data to by channel
        for(i in 1:nrow(dat)){
                print(i)
                dat.row <- dat[i, ]
                runInfo <- filter(dat.info, Runs == dat.row$Experiment & 
                        Channel == dat.row$Channel)
                runInfo$Channel <- NULL # remove redundant column
                tmp <- cbind(dat.row, runInfo)
                df <- rbind(df, tmp)
        }
        
        # remove any NAs, Blank runs, and thermal rings
        df <- df[complete.cases(df),]
        df <- filter(df, Treatment != "Blank" & Target != "thermal")
        
        # remove reduntant columns
        df$run <- NULL
        df$Step <- NULL
        df$Runs <- NULL
        df$Channel <- NULL
        
        # saves data
        write_csv(df, "compiledLabeled.csv")
}


CompileAndProcess <- function(){
        GetName()
        CompileData()
        CleanData()
}
